packages/pos_tagger/android/src/main/java/com/example/pos_tagger/PosTaggerPlugin.java
package com.example.pos_tagger;

import androidx.annotation.NonNull;

import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;

import io.flutter.embedding.engine.plugins.FlutterPlugin;
import io.flutter.plugin.common.MethodCall;
import io.flutter.plugin.common.MethodChannel;
import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
import io.flutter.plugin.common.MethodChannel.Result;

/** PosTaggerPlugin */
public class PosTaggerPlugin implements FlutterPlugin, MethodCallHandler {
  /// The MethodChannel that will the communication between Flutter and native
  /// Android
  ///
  /// This local reference serves to register the plugin with the Flutter Engine
  /// and unregister it
  /// when the Flutter Engine is detached from the Activity
  private MethodChannel channel;

  // Create a PosTagger object
  private final PosTagger posTagger = PosTagger.getInstance();


  @Override
  public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
    channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "pos_tagger");
    channel.setMethodCallHandler(this);
  }

  @Override
  public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
    if (call.method.equals("getPlatformVersion")) {
      result.success("Android " + android.os.Build.VERSION.RELEASE);
    } else if (call.method.equals("splitter")) {
      String input = call.argument("input");
      if (input != null) {
        List<String> output = new ArrayList<>(Arrays.asList(input.split("")));
        result.success(output);
      } else {
        result.error("INVALID_INPUT", "Input is null", null);
      }
    } else if (call.method.equals("posTag")) {
      String input = call.argument("input");
      if (input != null) {
        posTagger.evaluate(input);
        result.success("POS tagging completed");
      } else {
        result.error("INVALID_INPUT", "Input is null", null);
      }
    } else {
      result.notImplemented();
    }
  }

  @Override
  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
    channel.setMethodCallHandler(null);
  }
}
